Correção De Classes e aplicação do GenericDAO
Aplicação do Método salvar no generic DAO, e classe de teste de EstadoDAOTets, e correção de tipos primitivos nas classes

+++ src/main/java/com/pro/alysson/primeiroprojeto/domain/GenericDomain.java
@@ -12,13 +12,13 @@ import javax.persistence.MappedSuperclass;
 public class GenericDomain implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long codigo;
+	private int codigo;
 
-	public Long getCodigo() {
+	public int getCodigo() {
 		return codigo;
 	}
 
-	public void setCodigo(Long codigo) {
+	public void setCodigo(int codigo) {
 		this.codigo = codigo;
 	}
 }
